Adding total_value to cart model

+++ API/models/product_model.py
@@ -5,7 +5,7 @@ class Product(db.Model):
     product_name = db.Column(db.String(100), nullable=False)
     description = db.Column(db.String(100), nullable=False) 
     quantity = db.Column(db.Integer,nullable=False)
-    regular_price = db.Column(db.Float, nullable=False)
+    regular_price = db.Column(db.Numeric(10,2), nullable=False)
     
     def __repr__(self):
         return f"Product('{self.id}', '{self.product_name}', '{self.description}', '{self.quantity}','{self.regular_price}')"
